requirements.txt
psaw==0.0.12
progressbar2==3.51.3